fix: profile card: move StateStatus to separate files
