+++ src/components/Headers/ModerationActionHeader/ModerationActionHeader.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { Button } from '@/components/Button/Button'
 import GoBackButton from '@/components/GoBackButton/GoBackButton'
-import { StateStatus } from '@/components/ProfileList/ModerationProfileListItem'
+import { StateStatus } from '@/components/StateStatus/StateStatus'
 import { useModal } from '@/contexts/ModalContext'
 
 import AssignRoleModal from '@/components/AssignRoleModal/AssignRoleModal'
